using provider and useSelector
Provider is used to connect app with redux store -> for that we need to provide that store to react app. Access to redux will have only components that are wrapped with <Provider>

import { useStore } gives access to the store, but { useSelector } allows us to automatically select a part of our state managed by the store. useSelector() determines which piece of data we wanna extract from our store, parameters: (function receive the state managed by redux => returns the part of state which we wanna extract). Redux automatically sets up a subscription to the store for this component -> that means the component will be automatically updated and receive the latest value whenever redux store change.