refactoring code using useRouteMatch
useRouteMatch - if we change router path in app.js, we must to change it in nested roots too, but this hook helps us with that.

location.pathname gives the URL of the current page, while match.url gives the path you gave to the <Route> that is rendering the page. For example, while visiting /quotes/something, inside AllQuotes we will have /quotes/something as location.pathname, while match.url will be just /quotes because in the Route wrapping AllQuotes we have path='/quotes'.